Updated DB for chat history, sessions
