--- backend/SemanticKernel.AIAgentBackend/SemanticKernel.AIAgentBackend/Controllers/UserManagementController.cs
@@ -1,99 +1,0 @@
-﻿using Google;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using SemanticKernel.AIAgentBackend.Data;
-using System.Security.Claims;
-
-namespace SemanticKernel.AIAgentBackend.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    [Authorize]
-    public class UserManagementController : ControllerBase
-    {
-        private readonly AppDbContext _context;
-
-        public UserManagementController(AppDbContext context)
-        {
-            _context = context;
-        }
-
-        [HttpGet("profile")]
-        public async Task<IActionResult> GetProfile()
-        {
-            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            if (string.IsNullOrEmpty(userId))
-                return Unauthorized();
-
-            var user = await _context.Users
-                .Include(u => u.UserRoles)
-                .ThenInclude(ur => ur.Role)
-                .FirstOrDefaultAsync(u => u.Id == Guid.Parse(userId));
-
-            if (user == null)
-                return NotFound();
-
-            return Ok(new
-            {
-                user.Id,
-                user.Email,
-                user.FirstName,
-                user.LastName,
-                Roles = user.UserRoles.Select(ur => ur.Role.Name).ToList()
-            });
-        }
-
-        [HttpGet]
-        [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> GetAllUsers()
-        {
-            var users = await _context.Users
-                .Include(u => u.UserRoles)
-                .ThenInclude(ur => ur.Role)
-                .Select(u => new
-                {
-                    u.Id,
-                    u.Email,
-                    u.FirstName,
-                    u.LastName,
-                    u.CreatedAt,
-                    u.LastLoginAt,
-                    u.IsActive,
-                    Roles = u.UserRoles.Select(ur => ur.Role.Name).ToList()
-                })
-                .ToListAsync();
-
-            return Ok(users);
-        }
-
-        [HttpPut("{id}/roles")]
-        [Authorize(Policy = "RequireAdminRole")]
-        public async Task<IActionResult> UpdateUserRoles(Guid id, [FromBody] List<string> roleNames)
-        {
-            var user = await _context.Users
-                .Include(u => u.UserRoles)
-                .FirstOrDefaultAsync(u => u.Id == id);
-
-            if (user == null)
-                return NotFound();
-
-            // Remove existing roles
-            _context.UserRoles.RemoveRange(user.UserRoles);
-
-            // Add new roles
-            var roles = await _context.Roles
-                .Where(r => roleNames.Contains(r.Name))
-                .ToListAsync();
-
-            foreach (var role in roles)
-            {
-                user.UserRoles.Add(new() { UserId = user.Id, RoleId = role.Id });
-            }
-
-            await _context.SaveChangesAsync();
-            return NoContent();
-        }
-    }
-}
