--- backend/SemanticKernel.AIAgentBackend/SemanticKernel.AIAgentBackend/Models/Domain/ApplicationRole.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace SemanticKernel.AIAgentBackend.Models.Domain
-{
-    public class ApplicationRole : IdentityRole
-    {
-        public string? Description { get; set; }
-        public DateTime CreatedAt { get; set; }
-    }
-}
