+++ backend/SemanticKernel.AIAgentBackend/SemanticKernel.AIAgentBackend/Models/Domain/ChatHistory.cs
@@ -5,6 +5,10 @@
         public Guid Id { get; set; }
 
         public Guid SessionId { get; set; }
+        public SessionSummary SessionSummary { get; set; } = null!;
+
+        public Guid UserId { get; set; }
+        public User User { get; set; } = null!;
 
         public string? Message { get; set; }
